Clean up core.py further to just basics
